Fix typing of pixel_coords_to_space_coords

+++ manim/renderer/opengl_renderer.py
@@ -45,6 +45,7 @@ if TYPE_CHECKING:
     from manim.animation.animation import Animation
     from manim.mobject.mobject import Mobject
     from manim.scene.scene import Scene
+    from manim.typing import Point3D
 
 
 __all__ = ["OpenGLCamera", "OpenGLRenderer"]
@@ -587,7 +588,7 @@ class OpenGLRenderer:
     # top_left flag should be set to True when using a GUI framework
     # where the (0,0) is at the top left: e.g. PySide6
     def pixel_coords_to_space_coords(
-        self, px: int, py: int, relative: bool = False, top_left: bool = False
+        self, px: float, py: float, relative: bool = False, top_left: bool = False
     ) -> Point3D:
         pixel_shape = self.get_pixel_shape()
         if pixel_shape is None:
